+++ config/config.go
@@ -3,6 +3,7 @@ package config
 import (
 	"crypto/tls"
 	"crypto/x509"
+	"runtime"
 	"time"
 
 	"google.golang.org/grpc"
@@ -283,8 +284,16 @@ func New(opts ...Option) Config {
 }
 
 func defaults() (c *config) {
-	certPool, err := x509.SystemCertPool()
-	if err != nil {
+	var (
+		certPool *x509.CertPool
+		err      error
+	)
+	if runtime.GOOS != "darwin" {
+		certPool, err = x509.SystemCertPool()
+		if err != nil {
+			certPool = x509.NewCertPool()
+		}
+	} else {
 		certPool = x509.NewCertPool()
 	}
 	return &config{