+++ test/ratelimiter_test.go
@@ -5,6 +5,7 @@ package test
 
 import (
 	"context"
+	"log"
 	"os"
 	"testing"
 	"time"
@@ -16,12 +17,29 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
 	internal "github.com/ydb-platform/ydb-go-sdk/v3/internal/ratelimiter"
 	public "github.com/ydb-platform/ydb-go-sdk/v3/ratelimiter"
+	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
 const (
 	testCoordinationNodePath = "/local/test"
 	testResource             = "test_res"
 )
+
+func driverTrace() trace.Driver {
+	var t trace.Driver
+	trace.Stub(&t, func(name string, args ...interface{}) {
+		log.Printf("[driver] %s: %+v", name, trace.ClearContext(args))
+	})
+	return t
+}
+
+func tableTrace() trace.Table {
+	var t trace.Table
+	trace.Stub(&t, func(name string, args ...interface{}) {
+		log.Printf("[table] %s: %+v", name, trace.ClearContext(args))
+	})
+	return t
+}
 
 func TestRateLimiter(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*2)